fi-mcp-project/flask-backend/requirements.txt
flask
requests
flask-cors
google-genai
python-dotenv 